requirements.txt
Django==2.1.2
django-extensions==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-rest-passwordreset
